cleaned up code, getting it ready to remove child component
