+++ client/src/components/Inventory/Inventory.jsx
@@ -5,6 +5,7 @@ import { InventoryContext } from '../../contexts/inventory.context';
 import { getInventoryList } from '../../services/inventoryAPIcalls' // can be used instead of context
 import { createInventoryItem } from '../../services/inventoryAPIcalls'
 
+import AddProductRow from './AddProductRow';
 import SettingsPopup from './popups/settingsPopup';
 import OrderNowPopup from './popups/orderNowPopup';
 import IncomingPopup from './popups/incomingPopup';
@@ -17,61 +18,16 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 export default function Inventory() {
   const { inventory } = useContext(InventoryContext);
   const [popup, setPopup] = useState(null);
-  const [addProdRow, setAddProdRow] = useState([]);
-  const [addProdInfo, setAddProdInfo] = useState({
-    sku: '',
-    brand: '',
-    productName: '',
-    description: '',
-    inStock: '',
-    reorderAt: '',
-    orderQty: '',
-    unitPrice: ''
-  });
-
-  // ---------------- add product functions start ---------------------
-  const handleAddProd_InputChange = (e) => {
-    // (e) represents the event - the exact input that changed
-    const { name, value } = e.target;
-    setAddProdInfo((prevState) => ({ ...prevState, [name]: value }));
-  };
-
-  async function handleCreateItem(e) {
-    e.preventDefault();
-    const response = await createInventoryItem(addProdInfo)
-    console.log(response)
-    // clear fields after response succeeds
-    clearProdInputFields()
-  }
-
-  function clearProdInputFields() {
-    setAddProdInfo(prevState => {
-      return Object.fromEntries(Object.keys(prevState).map(key => [key, '']));
-    });
-  }
-  // ---------------- add product functions end  ---------------------
 
   const handleClick = (event) => {
     setPopup(event.target.id);
   };
 
-  const [tableHeader, setTableHeader] = useState([
-    "SKU",
-    "Brand",
-    "Name",
-    "Description",
-    "In Stock",
-    "Reorder At",
-    "Order QTY",
-    "Incoming Orders",
-    "Order Now",
-    "Settings",
+  const [tableHeader, setTableHeader] = useState(["SKU", "Brand", "Name", "Description", "In Stock", "Reorder At", "Order QTY", "Incoming Orders", "Order Now", "Settings",
   ]);
-
 
   const [rows, setRows] = useState([]);
   const [rowAdded, setRowAdded] = useState(false);
-
 
   // changing name from displayRow to handleDisplayRow
   const handleDisplayRow = () => {
@@ -84,18 +40,7 @@ export default function Inventory() {
   };
 
   const handleHeaderChange = (newHeader, reset = false) => {
-    const defaultHeader = [
-      "SKU",
-      "Brand",
-      "Name",
-      "Description",
-      "In Stock",
-      "Reorder At",
-      "Order QTY",
-      "Incoming Orders",
-      "Order Now",
-      "Settings",
-    ];
+    const defaultHeader = ["SKU", "Brand", "Name", "Description", "In Stock", "Reorder At", "Order QTY", "Incoming Orders", "Order Now", "Settings",];
     reset ? setTableHeader(defaultHeader) : setTableHeader(newHeader);
   };
 
@@ -118,91 +63,36 @@ export default function Inventory() {
         </thead>
         <tbody className="inventory-items-container">
 
-          {rowAdded && (
-            <tr>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs sku"
-                  name="sku"
-                  value={addProdInfo.sku}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs brand"
-                  name="brand"
-                  value={addProdInfo.brand}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs name"
-                  name="productName"
-                  value={addProdInfo.productName}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs desc"
-                  name="description"
-                  value={addProdInfo.description}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs"
-                  name="inStock"
-                  value={addProdInfo.inStock}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs"
-                  name="reorderAt"
-                  value={addProdInfo.reorderAt}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs"
-                  name="orderQty"
-                  value={addProdInfo.orderQty}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <input
-                  type="text" className="dynamic-inputs unit-price"
-                  name="unitPrice"
-                  value={addProdInfo.unitPrice}
-                  onChange={handleAddProd_InputChange} />
-              </td>
-              <td>
-                <form onSubmit={handleCreateItem}>
-                  <button type="submit">Save</button>
-                </form>
-              </td>
-              <td>
-                <button onClick={() => {
-                  handleHideRow()
-                  handleHeaderChange(null, true);
-                  clearProdInputFields()
-                }
-                }>Cancel</button>
-              </td>
-            </tr>
-          )}
-          {/* ))} */}
+          <AddProductRow
+            rowAdded={rowAdded}
+            handleHideRow={handleHideRow}
+            handleHeaderChange={handleHeaderChange}
+          />
+
           {inventory.map((item) => (
+            // use key here to get specific item to get (for popup) update or delete
             <tr key={item.id}>
-              <td>{item.sku}</td>
-              <td>{item.brand}</td>
-              <td className="item-name">{item.productName}</td>
-              <td className="item-description">{item.description}</td>
-              <td>{item.inStock}</td>
+              <td>
+                {item.sku}
+              </td>
+              <td>
+                {item.brand}
+              </td>
+              <td
+                className="item-name">
+                {item.productName}
+              </td>
+              <td
+                className="item-description">
+                {item.description}
+              </td>
+              <td>
+                {item.inStock}
+              </td>
               <td>
                 <input
                   className="dynamic-inputs"
-                  id="name-input"
+                  id="reorderAt"
                   type="text"
                   defaultValue={item.reorderAt}
                 // onChange={handleOrderQtyChange}
@@ -211,7 +101,7 @@ export default function Inventory() {
               <td>
                 <input
                   className="dynamic-inputs"
-                  id="name-input"
+                  id="orderQty"
                   type="text"
                   defaultValue={item.orderQty}
                 // onChange={handleOrderQtyChange}
